@@ -1,25 +1,27 @@
 [project]
 name = "curl_cffi"
-version = "0.14.0b2"
+version = "0.15.0b4"
 authors = [{ name = "lexiforest", email = "infinitesheldon@gmail.com" }]
 description = "libcurl ffi bindings for Python, with impersonation support."
-license = { text = "MIT License" }
+license = "MIT"
+license-files = ["LICENSE"]
 dependencies = [
-    "cffi>=1.12.0",
+    "cffi>=2.0.0",
     "certifi>=2024.2.2",
 ]
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 urls = { "repository" = "https://github.com/lexiforest/curl_cffi" }
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Developers",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+    "Programming Language :: Python :: Free Threading",
 ]
 
 
