@@ -12,8 +12,7 @@ print(reply)
 
 
 async def async_examples():
-    async with curl_cffi.AsyncSession() as s:
-        ws = await s.ws_connect(URL)
+    async with curl_cffi.AsyncSession() as s, s.ws_connect(URL) as ws:
         await ws.send(b"Bar")
         reply = await ws.recv()
         print(reply)
