@@ -3,6 +3,7 @@ import os
 import platform
 import shutil
 import struct
+import sys
 import tempfile
 from glob import glob
 from pathlib import Path
@@ -11,7 +12,17 @@ from urllib.request import urlretrieve
 from cffi import FFI
 
 # this is the upstream libcurl-impersonate version
-__version__ = "1.2.2"
+__version__ = "1.4.2"
+
+
+def is_android_env() -> bool:
+    return bool(
+        sys.platform == "android"
+        or os.environ.get("CIBW_PLATFORM") == "android"
+        or os.environ.get("ANDROID_ROOT")
+        or os.environ.get("ANDROID_DATA")
+        or os.environ.get("TERMUX_VERSION")
+    )
 
 
 def detect_arch():
@@ -19,24 +30,26 @@ def detect_arch():
         archs = json.loads(f.read())
 
     uname = platform.uname()
+    uname_system = "Android" if is_android_env() else uname.system
     glibc_flavor = "gnueabihf" if uname.machine in ["armv7l", "armv6l"] else "gnu"
 
     libc, _ = platform.libc_ver()
     # https://github.com/python/cpython/issues/87414
     libc = glibc_flavor if libc == "glibc" else "musl"
+    if is_android_env():
+        libc = "android"
     pointer_size = struct.calcsize("P") * 8
 
     for arch in archs:
         if (
-            arch["system"] == uname.system
+            arch["system"] == uname_system
             and arch["machine"] == uname.machine
             and arch["pointer_size"] == pointer_size
             and ("libc" not in arch or arch.get("libc") == libc)
         ):
-            if arch["libdir"]:
+            if arch.get("libdir"):
                 arch["libdir"] = os.path.expanduser(arch["libdir"])
             else:
-                global tmpdir
                 if "CI" in os.environ:
                     tmpdir = "./tmplibdir"
                     os.makedirs(tmpdir, exist_ok=True)
@@ -49,12 +62,18 @@ def detect_arch():
 
 
 arch = detect_arch()
-print(f"Using {arch['libdir']} to store libcurl-impersonate")
+link_type = arch.get("link_type")
+libdir = Path(arch["libdir"])
+is_static = link_type == "static"
+is_dynamic = link_type == "dynamic"
+is_android = arch.get("libc") == "android"
+print(f"Using {libdir} to store libcurl-impersonate")
 
 
 def download_libcurl():
-    if (Path(arch["libdir"]) / arch["so_name"]).exists():
-        print(".so files already downloaded.")
+    expected = libdir / arch["obj_name"]
+    if expected.exists():
+        print(f"libcurl-impersonate: {expected} already downloaded.")
         return
 
     file = "libcurl-impersonate.tar.gz"
@@ -63,47 +82,33 @@ def download_libcurl():
     url = (
         f"https://github.com/lexiforest/curl-impersonate/releases/download/"
         f"v{__version__}/libcurl-impersonate-v{__version__}"
-        f".{arch['so_arch']}-{sysname}.tar.gz"
+        f".{arch['arch']}-{sysname}.tar.gz"
     )
 
     print(f"Downloading libcurl-impersonate from {url}...")
     urlretrieve(url, file)
 
     print("Unpacking downloaded files...")
-    os.makedirs(arch["libdir"], exist_ok=True)
-    shutil.unpack_archive(file, arch["libdir"])
+    os.makedirs(libdir, exist_ok=True)
+    shutil.unpack_archive(file, libdir)
 
     if arch["system"] == "Windows":
-        for file in glob(os.path.join(arch["libdir"], "lib/*.lib")):
-            shutil.move(file, arch["libdir"])
-        for file in glob(os.path.join(arch["libdir"], "bin/*.dll")):
-            shutil.move(file, arch["libdir"])
+        for file in glob(str(libdir / "lib/*.lib")):
+            shutil.move(file, libdir)
+        for file in glob(str(libdir / "bin/*.dll")):
+            shutil.move(file, libdir)
 
-    print("Files after unpacking")
-    print(os.listdir(arch["libdir"]))
+    print("Files after unpacking:")
+    print(os.listdir(libdir))
 
 
 def get_curl_archives():
-    print("Files for linking")
-    print(os.listdir(arch["libdir"]))
-    if arch["system"] == "Linux" and arch.get("link_type") == "static":
+    print("Files in linking directory:")
+    print(os.listdir(libdir))
+    if is_static:
         # note that the order of libraries matters
         # https://stackoverflow.com/a/36581865
-        return [
-            f"{arch['libdir']}/libcurl-impersonate.a",
-            f"{arch['libdir']}/libssl.a",
-            f"{arch['libdir']}/libcrypto.a",
-            f"{arch['libdir']}/libz.a",
-            f"{arch['libdir']}/libzstd.a",
-            f"{arch['libdir']}/libnghttp2.a",
-            f"{arch['libdir']}/libngtcp2.a",
-            f"{arch['libdir']}/libngtcp2_crypto_boringssl.a",
-            f"{arch['libdir']}/libnghttp3.a",
-            f"{arch['libdir']}/libbrotlidec.a",
-            f"{arch['libdir']}/libbrotlienc.a",
-            f"{arch['libdir']}/libbrotlicommon.a",
-            f"{arch['libdir']}/libcares.a",
-        ]
+        return [str(libdir / arch["obj_name"])]
     else:
         return []
 
@@ -128,11 +133,8 @@ def get_curl_libraries():
             "brotlidec",
             "brotlicommon",
             "iphlpapi",
-            "cares",
         ]
-    elif arch["system"] == "Darwin" or (
-        arch["system"] == "Linux" and arch.get("link_type") == "dynamic"
-    ):
+    elif is_dynamic:
         return ["curl-impersonate"]
     else:
         return []
@@ -143,21 +145,39 @@ system = platform.system()
 root_dir = Path(__file__).parent.parent
 download_libcurl()
 
+# With mega archive, we only have one to link
+static_libs = get_curl_archives()
+extra_link_args = []
+if is_static:
+    if system == "Darwin":
+        extra_link_args = [
+            f"-Wl,-force_load,{static_libs[0]}",
+            "-lc++",
+        ]
+    elif system in ("Linux", "Android"):
+        cxx_lib = "-lc++" if is_android else "-lstdc++"
+        extra_link_args = [
+            "-Wl,--whole-archive",
+            static_libs[0],
+            "-Wl,--no-whole-archive",
+            cxx_lib,
+        ]
+
+libraries = get_curl_libraries()
 
 ffibuilder.set_source(
     "curl_cffi._wrapper",
     """
         #include "shim.h"
     """,
-    # FIXME from `curl-impersonate`
+    library_dirs=[str(libdir)],
     libraries=get_curl_libraries(),
-    extra_objects=get_curl_archives(),
-    library_dirs=[arch["libdir"]],
+    extra_objects=[],  # linked via extra_link_args
     source_extension=".c",
     include_dirs=[
         str(root_dir / "include"),
         str(root_dir / "ffi"),
-        str(Path(arch["libdir"]) / "include"),
+        str(libdir / "include"),
     ],
     sources=[
         str(root_dir / "ffi/shim.c"),
@@ -165,7 +185,7 @@ ffibuilder.set_source(
     extra_compile_args=(
         ["-Wno-implicit-function-declaration"] if system == "Darwin" else []
     ),
-    extra_link_args=(["-lstdc++"] if system != "Windows" else []),
+    extra_link_args=extra_link_args,
 )
 
 with open(root_dir / "ffi/cdef.c") as f:
